# artifacts/api-server/src/routes/experience.ts
import { Router } from "express";
import { db } from "@workspace/db";
import { experienceTable } from "@workspace/db";
import { eq, asc } from "drizzle-orm";

const router = Router();

router.get("/", async (req, res) => {
  try {
    const rows = await db.select().from(experienceTable).orderBy(asc(experienceTable.sortOrder), asc(experienceTable.id));
    return res.json(rows);
  } catch (err) {
    req.log.error({ err }, "Failed to list experience");
    return res.status(500).json({ error: "Internal server error" });
  }
});

router.post("/", async (req, res) => {
  try {
    const inserted = await db.insert(experienceTable).values(req.body).returning();
    return res.status(201).json(inserted[0]);
  } catch (err) {
    req.log.error({ err }, "Failed to create experience");
    return res.status(500).json({ error: "Internal server error" });
  }
});

router.patch("/:id", async (req, res) => {
  try {
    const id = parseInt(req.params.id, 10);
    const updated = await db.update(experienceTable).set(req.body).where(eq(experienceTable.id, id)).returning();
    if (!updated.length) return res.status(404).json({ error: "Not found" });
    return res.json(updated[0]);
  } catch (err) {
    req.log.error({ err }, "Failed to update experience");
    return res.status(500).json({ error: "Internal server error" });
  }
});

router.delete("/:id", async (req, res) => {
  try {
    const id = parseInt(req.params.id, 10);
    await db.delete(experienceTable).where(eq(experienceTable.id, id));
    return res.status(204).send();
  } catch (err) {
    req.log.error({ err }, "Failed to delete experience");
    return res.status(500).json({ error: "Internal server error" });
  }
});

export default router;