# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Portfolio website API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Hero {
  id: number;
  name: string;
  title: string;
  subtitle: string;
  bio: string;
  ctaPrimary: string;
  ctaSecondary: string;
  avatarUrl: string;
  /** @nullable */
  resumeUrl: string | null;
  updatedAt?: string;
}

export interface HeroInput {
  name: string;
  title: string;
  subtitle: string;
  bio: string;
  ctaPrimary: string;
  ctaSecondary: string;
  avatarUrl: string;
  resumeUrl?: string;
}

export interface About {
  id: number;
  bio: string;
  profileImageUrl: string;
  location: string;
  email: string;
  /** @nullable */
  phone: string | null;
  yearsExperience: number;
  projectsCompleted: number;
  happyClients: number;
  updatedAt?: string;
}

export interface AboutInput {
  bio: string;
  profileImageUrl: string;
  location: string;
  email: string;
  phone?: string;
  yearsExperience: number;
  projectsCompleted: number;
  happyClients: number;
}

export interface Skill {
  id: number;
  name: string;
  category: string;
  level: number;
  /** @nullable */
  icon: string | null;
  sortOrder: number;
  createdAt?: string;
}

export interface SkillInput {
  name: string;
  category: string;
  level: number;
  icon?: string;
  sortOrder?: number;
}

export interface SkillUpdate {
  name?: string;
  category?: string;
  level?: number;
  icon?: string;
  sortOrder?: number;
}

export interface Project {
  id: number;
  title: string;
  description: string;
  /** @nullable */
  longDescription?: string | null;
  technologies: string[];
  /** @nullable */
  githubUrl?: string | null;
  /** @nullable */
  liveUrl?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  featured: boolean;
  category: string;
  sortOrder: number;
  createdAt: string;
}

export interface ProjectInput {
  title: string;
  description: string;
  longDescription?: string;
  technologies: string[];
  githubUrl?: string;
  liveUrl?: string;
  imageUrl?: string;
  featured?: boolean;
  category: string;
  sortOrder?: number;
}

export interface ProjectUpdate {
  title?: string;
  description?: string;
  longDescription?: string;
  technologies?: string[];
  githubUrl?: string;
  liveUrl?: string;
  imageUrl?: string;
  featured?: boolean;
  category?: string;
  sortOrder?: number;
}

export interface Service {
  id: number;
  icon: string;
  title: string;
  description: string;
  /** @nullable */
  pricing?: string | null;
  features?: string[];
  sortOrder: number;
  createdAt: string;
}

export interface ServiceInput {
  icon: string;
  title: string;
  description: string;
  pricing?: string;
  features?: string[];
  sortOrder?: number;
}

export interface ServiceUpdate {
  icon?: string;
  title?: string;
  description?: string;
  pricing?: string;
  features?: string[];
  sortOrder?: number;
}

export interface Experience {
  id: number;
  company: string;
  position: string;
  startDate: string;
  /** @nullable */
  endDate: string | null;
  description: string;
  current: boolean;
  /** @nullable */
  companyUrl?: string | null;
  /** @nullable */
  logoUrl?: string | null;
  sortOrder: number;
  createdAt?: string;
}

export interface ExperienceInput {
  company: string;
  position: string;
  startDate: string;
  endDate?: string;
  description: string;
  current?: boolean;
  companyUrl?: string;
  logoUrl?: string;
  sortOrder?: number;
}

export interface ExperienceUpdate {
  company?: string;
  position?: string;
  startDate?: string;
  endDate?: string;
  description?: string;
  current?: boolean;
  companyUrl?: string;
  logoUrl?: string;
  sortOrder?: number;
}

export interface Education {
  id: number;
  institution: string;
  degree: string;
  field: string;
  startYear: number;
  /** @nullable */
  endYear: number | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  logoUrl?: string | null;
  sortOrder: number;
  createdAt?: string;
}

export interface EducationInput {
  institution: string;
  degree: string;
  field: string;
  startYear: number;
  endYear?: number;
  description?: string;
  logoUrl?: string;
  sortOrder?: number;
}

export interface EducationUpdate {
  institution?: string;
  degree?: string;
  field?: string;
  startYear?: number;
  endYear?: number;
  description?: string;
  logoUrl?: string;
  sortOrder?: number;
}

export interface Testimonial {
  id: number;
  name: string;
  role: string;
  company: string;
  quote: string;
  /** @nullable */
  avatarUrl?: string | null;
  rating?: number;
  sortOrder: number;
  createdAt?: string;
}

export interface TestimonialInput {
  name: string;
  role: string;
  company: string;
  quote: string;
  avatarUrl?: string;
  rating?: number;
  sortOrder?: number;
}

export interface TestimonialUpdate {
  name?: string;
  role?: string;
  company?: string;
  quote?: string;
  avatarUrl?: string;
  rating?: number;
  sortOrder?: number;
}

export interface BlogPost {
  id: number;
  title: string;
  slug: string;
  content: string;
  excerpt: string;
  /** @nullable */
  coverImageUrl?: string | null;
  published: boolean;
  featured: boolean;
  /** @nullable */
  category?: string | null;
  tags?: string[];
  /** @nullable */
  readTime?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface BlogPostInput {
  title: string;
  slug: string;
  content: string;
  excerpt: string;
  coverImageUrl?: string;
  published?: boolean;
  featured?: boolean;
  category?: string;
  tags?: string[];
  readTime?: number;
}

export interface BlogPostUpdate {
  title?: string;
  slug?: string;
  content?: string;
  excerpt?: string;
  coverImageUrl?: string;
  published?: boolean;
  featured?: boolean;
  category?: string;
  tags?: string[];
  readTime?: number;
}

export interface ContactMessage {
  id: number;
  name: string;
  email: string;
  subject: string;
  message: string;
  read: boolean;
  replied?: boolean;
  createdAt: string;
}

export interface ContactMessageInput {
  name: string;
  email: string;
  subject: string;
  message: string;
}

export interface ContactMessageUpdate {
  read?: boolean;
  replied?: boolean;
}

export interface SiteSettings {
  id: number;
  siteTitle: string;
  siteDescription: string;
  siteUrl: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  githubUrl?: string | null;
  /** @nullable */
  linkedinUrl?: string | null;
  /** @nullable */
  twitterUrl?: string | null;
  /** @nullable */
  instagramUrl?: string | null;
  /** @nullable */
  youtubeUrl?: string | null;
  /** @nullable */
  faviconUrl?: string | null;
  /** @nullable */
  ogImageUrl?: string | null;
  /** @nullable */
  googleAnalyticsId?: string | null;
  updatedAt?: string;
}

export interface SiteSettingsInput {
  siteTitle: string;
  siteDescription: string;
  siteUrl: string;
  email: string;
  phone?: string;
  address?: string;
  githubUrl?: string;
  linkedinUrl?: string;
  twitterUrl?: string;
  instagramUrl?: string;
  youtubeUrl?: string;
  faviconUrl?: string;
  ogImageUrl?: string;
  googleAnalyticsId?: string;
}

export interface DashboardStats {
  totalProjects: number;
  totalBlogPosts: number;
  totalMessages: number;
  unreadMessages: number;
  totalSkills: number;
  totalTestimonials: number;
  totalServices: number;
  totalExperience: number;
}

export type ListProjectsParams = {
category?: string;
featured?: boolean;
};

export type ListBlogPostsParams = {
category?: string;
featured?: boolean;
};