# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Portfolio website API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a single project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # === SERVICES ===
...
  /experience:
    get:
      operationId: listExperience
      tags: [experience]
      summary: List all experience entries
      responses:
        "200":
          description: List of experience
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Experience"
    post:
      operationId: createExperience
      tags: [experience]
      summary: Create an experience entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExperienceInput"
      responses:
        "201":
          description: Created experience
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Experience"
  /experience/{id}:
    patch:
      operationId: updateExperience
      tags: [experience]
      summary: Update an experience entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExperienceUpdate"
      responses:
        "200":
          description: Updated experience
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Experience"
    delete:
      operationId: deleteExperience
      tags: [experience]
      summary: Delete an experience entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # === EDUCATION ===
  /education:
    get:
      operationId: listEducation
      tags: [education]
      summary: List all education entries
      responses:
        "200":
          description: List of education
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Education"
    post:
      operationId: createEducation
      tags: [education]
      summary: Create an education entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EducationInput"
      responses:
        "201":
          description: Created education
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Education"
  /education/{id}:
    patch:
      operationId: updateEducation
      tags: [education]
      summary: Update an education entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EducationUpdate"
      responses:
        "200":
          description: Updated education
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Education"
    delete:
      operationId: deleteEducation
      tags: [education]
      summary: Delete an education entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # === TESTIMONIALS ===
...
  /blog:
    get:
      operationId: listBlogPosts
      tags: [blog]
      summary: List all blog posts
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of blog posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlogPost"
    post:
      operationId: createBlogPost
      tags: [blog]
      summary: Create a blog post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlogPostInput"
      responses:
        "201":
          description: Created blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
  /blog/{id}:
    get:
      operationId: getBlogPost
      tags: [blog]
      summary: Get a single blog post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
    patch:
      operationId: updateBlogPost
      tags: [blog]
      summary: Update a blog post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlogPostUpdate"
      responses:
        "200":
          description: Updated blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
    delete:
      operationId: deleteBlogPost
      tags: [blog]
      summary: Delete a blog post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # === CONTACT ===
  /contact:
    get:
      operationId: listContactMessages
      tags: [contact]
      summary: List all contact messages (admin)
      responses:
        "200":
          description: List of contact messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContactMessage"
    post:
      operationId: sendContactMessage
      tags: [contact]
      summary: Submit a contact form message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactMessageInput"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactMessage"
  /contact/{id}:
    patch:
      operationId: updateContactMessage
      tags: [contact]
      summary: Mark message as read/replied
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactMessageUpdate"
      responses:
        "200":
          description: Updated message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactMessage"
    delete:
      operationId: deleteContactMessage
      tags: [contact]
      summary: Delete a contact message
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # === SETTINGS ===
...
components:
  schemas:
...
    Project:
      type: object
      required: [id, title, description, technologies, featured, category, sortOrder, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: string }
        longDescription: { type: ["string", "null"] }
        technologies: { type: array, items: { type: string } }
        githubUrl: { type: ["string", "null"] }
        liveUrl: { type: ["string", "null"] }
        imageUrl: { type: ["string", "null"] }
        featured: { type: boolean }
        category: { type: string }
        sortOrder: { type: integer }
        createdAt: { type: string }
    ProjectInput:
      type: object
      required: [title, description, technologies, category]
      properties:
        title: { type: string }
        description: { type: string }
        longDescription: { type: string }
        technologies: { type: array, items: { type: string } }
        githubUrl: { type: string }
        liveUrl: { type: string }
        imageUrl: { type: string }
        featured: { type: boolean }
        category: { type: string }
        sortOrder: { type: integer }
    ProjectUpdate:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        longDescription: { type: string }
        technologies: { type: array, items: { type: string } }
        githubUrl: { type: string }
        liveUrl: { type: string }
        imageUrl: { type: string }
        featured: { type: boolean }
        category: { type: string }
        sortOrder: { type: integer }

    # === SERVICE ===
...
    Experience:
      type: object
      required: [id, company, position, startDate, endDate, description, current, sortOrder]
      properties:
        id: { type: integer }
        company: { type: string }
        position: { type: string }
        startDate: { type: string }
        endDate: { type: ["string", "null"] }
        description: { type: string }
        current: { type: boolean }
        companyUrl: { type: ["string", "null"] }
        logoUrl: { type: ["string", "null"] }
        sortOrder: { type: integer }
        createdAt: { type: string }
    ExperienceInput:
      type: object
      required: [company, position, startDate, description]
      properties:
        company: { type: string }
        position: { type: string }
        startDate: { type: string }
        endDate: { type: string }
        description: { type: string }
        current: { type: boolean }
        companyUrl: { type: string }
        logoUrl: { type: string }
        sortOrder: { type: integer }
    ExperienceUpdate:
      type: object
      properties:
        company: { type: string }
        position: { type: string }
        startDate: { type: string }
        endDate: { type: string }
        description: { type: string }
        current: { type: boolean }
        companyUrl: { type: string }
        logoUrl: { type: string }
        sortOrder: { type: integer }

    # === EDUCATION ===
    Education:
      type: object
      required: [id, institution, degree, field, startYear, endYear, sortOrder]
      properties:
        id: { type: integer }
        institution: { type: string }
        degree: { type: string }
        field: { type: string }
        startYear: { type: integer }
        endYear: { type: ["integer", "null"] }
        description: { type: ["string", "null"] }
        logoUrl: { type: ["string", "null"] }
        sortOrder: { type: integer }
        createdAt: { type: string }
    EducationInput:
      type: object
      required: [institution, degree, field, startYear]
      properties:
        institution: { type: string }
        degree: { type: string }
        field: { type: string }
        startYear: { type: integer }
        endYear: { type: integer }
        description: { type: string }
        logoUrl: { type: string }
        sortOrder: { type: integer }
    EducationUpdate:
      type: object
      properties:
        institution: { type: string }
        degree: { type: string }
        field: { type: string }
        startYear: { type: integer }
        endYear: { type: integer }
        description: { type: string }
        logoUrl: { type: string }
        sortOrder: { type: integer }

    # === TESTIMONIAL ===
...
    BlogPost:
      type: object
      required: [id, title, slug, content, excerpt, published, featured, createdAt, updatedAt]
      properties:
        id: { type: integer }
        title: { type: string }
        slug: { type: string }
        content: { type: string }
        excerpt: { type: string }
        coverImageUrl: { type: ["string", "null"] }
        published: { type: boolean }
        featured: { type: boolean }
        category: { type: ["string", "null"] }
        tags: { type: array, items: { type: string } }
        readTime: { type: ["integer", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
    BlogPostInput:
      type: object
      required: [title, slug, content, excerpt]
      properties:
        title: { type: string }
        slug: { type: string }
        content: { type: string }
        excerpt: { type: string }
        coverImageUrl: { type: string }
        published: { type: boolean }
        featured: { type: boolean }
        category: { type: string }
        tags: { type: array, items: { type: string } }
        readTime: { type: integer }
    BlogPostUpdate:
      type: object
      properties:
        title: { type: string }
        slug: { type: string }
        content: { type: string }
        excerpt: { type: string }
        coverImageUrl: { type: string }
        published: { type: boolean }
        featured: { type: boolean }
        category: { type: string }
        tags: { type: array, items: { type: string } }
        readTime: { type: integer }

    # === CONTACT MESSAGE ===
    ContactMessage:
      type: object
      required: [id, name, email, subject, message, read, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        subject: { type: string }
        message: { type: string }
        read: { type: boolean }
        replied: { type: boolean }
        createdAt: { type: string }
    ContactMessageInput:
      type: object
      required: [name, email, subject, message]
      properties:
        name: { type: string }
        email: { type: string }
        subject: { type: string }
        message: { type: string }
    ContactMessageUpdate:
      type: object
      properties:
        read: { type: boolean }
        replied: { type: boolean }

    # === SITE SETTINGS ===
...