# lib/api-zod/src/generated/types/contactMessageInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Portfolio website API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ContactMessageInput {
  name: string;
  email: string;
  subject: string;
  message: string;
}