# lib/api-zod/src/generated/types/education.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Portfolio website API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Education {
  id: number;
  institution: string;
  degree: string;
  field: string;
  startYear: number;
  /** @nullable */
  endYear: number | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  logoUrl?: string | null;
  sortOrder: number;
  createdAt?: string;
}